CustomView getting deleted twice when Browser goes into Background
OnPause() function is called for Browser activity when browser goes
into backgroud. If Browser was playing video, associated video view
receives suspend event twice that was causing video not to start
where it was left when browser went to background. This behavour was
observed after one of the commit from AOSP. Reverting the commit.

Change-Id: I386c6c8b1683def83056337ffcafe74ad29b2b1f