Do not reattach tab on orientation change
When the orientation changes, there is no need to detach and reattach
the current tab. This was causing a crash when Flash forces an
orientation change to play fullscreen video. The previous fix stopped
RealViewSwitcher from calling onScreenSwitched if the screen doesn't
change after a scroll, however the Tab needs to know when the scroll
completes regardless of if the Tab has changed so that it can show the
URL bar.

Change-Id: Ib9bbaa0effddc7d38261739c666e0eed256c6b02
CRs-fixed: 329432