Prevent unnecessary view removal when using Slide Tab Transitions
When using the feature Slide Tab Transitions and a tab was set as
active, the tab view container was always removed and added again
to a RealViewSwitcher. Also, when a WebView is detached from a
window and it is in full screen mode, it dismisses the full screen
mode. This caused a problem when a plugin (e.g. Flash) wanted to
show something in fullscreen.

Now a tab view container will not be removed add added again to
the same RealViewSwitcher.

(cherry picked from commit 8704450b5c0f1bcdd7fa3b07b519052d1f34eb91)

Change-Id: Ic85f8cf93b69400783663a6b3171b0d896371eca