New Tab Animation
In stock Android, opening a new tab and navigating a link is not
distinguisable. This patch visually indicates that a new tab has been
created using an animation. Also fixes issue with keyboard briefly
showing during slide transitions and an issue with
attachTabToContentView being called twice for the same tab.

Change-Id: Ib80a9bddb6b3cf61e57f6ae12ec216481478da7e