Fix for orientation change crash
The RealViewSwitcher was getting its new tab view added out of bounds
after a landscape rotation occurred. This was because a tab was being
left behind after the orientation change. The code to stop reattachment
on orientation change was a bug fix for a crash that occurred when Flash
video embedded in the Browser went to fullscreen and changed the
orientation to landscape mode. However, the code changed to prevent
unnecessary reattachments, which should already handle this case, making
the original code for orientation change redundant and caused the loss
of a tab after changing orientation.

Change-Id: I174df8c514c95738452afc3c5a3292d736835a99
CRs-fixed: 386305